storage/requirements.txt
asgiref==3.8.1
Django==4.2.7
django-cors-headers==4.3.0
django-nested-admin==4.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pillow==10.1.0
python-monkey-business==1.1.0
sqlparse==0.5.3
gunicorn
python-dotenv==1.0.0
psycopg2-binary==2.9.9
django-filter==23.5
django-storages==1.14.2
boto3==1.34.14